test_requirements.txt
python-socketio==5.10.0
requests==2.31.0 